# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Todo API",
        "description": "This is an open and free to use API for developers to test their client projects.\n        All api calls require you to pass in <br> your API Key with the querystring\n        <b>?apikey=########-####-####-####-############</b>, <br><br>Please\n        register for your own API Key today at https://todoapi.net/register ",
        "version": "1.0"
    },
    "paths": {
...
        "/api/todos": {
            "get": {
                "tags": [
                    "Generic Todo"
                ],
                "summary": "Retrieve all todos",
                "operationId": "todo.index",
                "parameters": [
                    {
                        "name": "apikey",
                        "in": "query",
                        "description": "all api calls require the ?apikey querystring.",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "A collection of todos"
                    }
                }
            },
            "post": {
                "tags": [
                    "Generic Todo"
                ],
                "summary": "Create a new Todo.  There's a limit of 100 generic todos per developer account.",
                "operationId": "Todo.create",
                "parameters": [
                    {
                        "name": "apikey",
                        "in": "query",
                        "description": "all api calls require the ?apikey querystring.",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Submit a todo description and optionally a completion flag",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "description"
                                ],
                                "properties": {
                                    "description": {
                                        "type": "string"
                                    },
                                    "completed": {
                                        "type": "boolean"
                                    },
                                    "meta": {
                                        "type": "object"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Todo created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer"
                                        },
                                        "description": {
                                            "type": "string"
                                        },
                                        "completed": {
                                            "type": "boolean"
                                        },
                                        "meta": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/users/{user_id}/todos": {
            "get": {
                "tags": [
                    "User Todo"
                ],
                "summary": "Retrieve all todos",
                "operationId": "user.todo.index",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "path",
                        "description": "User Id",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "name": "apikey",
                        "in": "query",
                        "description": "A collection of todos for current user",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Collection of todos"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "User Todo"
                ],
                "summary": "Create a new Todo for current user.  There's a limit of 25 todos per user .",
                "operationId": "user.todo.create",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "path",
                        "description": "User Id",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "name": "apikey",
                        "in": "query",
                        "description": "all api calls require the ?apikey querystring.",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Submit a todo description and optionally a completion flag",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "description"
                                ],
                                "properties": {
                                    "description": {
                                        "type": "string"
                                    },
                                    "completed": {
                                        "type": "boolean"
                                    },
                                    "meta": {
                                        "type": "object"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Todo created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer"
                                        },
                                        "description": {
                                            "type": "string"
                                        },
                                        "completed": {
                                            "type": "boolean"
                                        },
                                        "meta": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/todos/{id}": {
            "get": {
                "tags": [
                    "Generic Todo"
                ],
                "summary": "Retrieve specific todo",
                "operationId": "todo.get",
                "parameters": [
                    {
                        "name": "apikey",
                        "in": "query",
                        "description": "all api calls require the ?apikey querystring.",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The id of the todo you wish to retrieve",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "A todo"
                    }
                }
            },
            "put": {
                "tags": [
                    "Generic Todo"
                ],
                "summary": "Create a new Todo",
                "operationId": "Todo.update",
                "parameters": [
                    {
                        "name": "apikey",
                        "in": "query",
                        "description": "all api calls require the ?apikey querystring.",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Todo ID",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Submit a todo description and optionally a completion flag",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "description"
                                ],
                                "properties": {
                                    "description": {
                                        "type": "string"
                                    },
                                    "completed": {
                                        "type": "boolean"
                                    },
                                    "meta": {
                                        "type": "object"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Todo updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer"
                                        },
                                        "description": {
                                            "type": "string"
                                        },
                                        "completed": {
                                            "type": "boolean"
                                        },
                                        "meta": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/todo/{id}": {
            "delete": {
                "tags": [
                    "Generic Todo"
                ],
                "summary": "Deletes the todo.",
                "operationId": "TodoController.Remove",
                "parameters": [
                    {
                        "name": "apikey",
                        "in": "query",
                        "description": "all api calls require the ?apikey querystring.",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Todo ID",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns a message"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "apiAuth": {
                "type": "http",
                "description": "Login with email and password to get the authentication token",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}